float-based math, apply TRS

+++ game-ai-practice/GraphicsEngine/Matrix3.h
@@ -5,31 +5,31 @@
 
 struct [[nodiscard]] mat3
 {
-	double element[3][3];
+	float element[3][3];
 
 	inline mat3() = default;
-	inline constexpr mat3(double val);
+	inline constexpr mat3(float val);
 	inline constexpr mat3(const vec3& col1, const vec3& col2, const vec3& col3);
 	 
-	inline constexpr double& operator()(int column, int row);
-	inline constexpr double operator()(int column, int row) const;
+	inline constexpr float& operator()(int column, int row);
+	inline constexpr float operator()(int column, int row) const;
 	 
 	inline constexpr mat3& operator+=(const mat3& rhs) noexcept;
 	inline constexpr mat3& operator-=(const mat3& rhs) noexcept;
-	inline constexpr mat3& operator*=(double rhs) noexcept;
-	inline constexpr mat3& operator/=(double rhs) noexcept;
+	inline constexpr mat3& operator*=(float rhs) noexcept;
+	inline constexpr mat3& operator/=(float rhs) noexcept;
 	inline constexpr mat3& operator*=(const mat3& rhs) noexcept;
 	 	 
 	inline constexpr mat3 operator+(const mat3& rhs) noexcept;
 	inline constexpr mat3 operator-(const mat3& rhs) noexcept;
-	inline constexpr mat3 operator*(double rhs) noexcept;
-	inline constexpr mat3 operator/(double rhs) noexcept;
+	inline constexpr mat3 operator*(float rhs) noexcept;
+	inline constexpr mat3 operator/(float rhs) noexcept;
 	inline constexpr mat3 operator*(const mat3& rhs) noexcept;
 	 
 	inline [[nodiscard]] constexpr bool operator==(const mat3& rhs) noexcept;
 	inline [[nodiscard]] constexpr bool operator!=(const mat3& rhs) noexcept;
 
-	inline double* GetElementPtr() { return &element[0][0]; }
+	inline float* GetElementPtr() { return &element[0][0]; }
 };
 
 namespace matrix3
